+++ packages/raystack/v1/components/emptystate/index.tsx
@@ -1,1 +1,1 @@
-export { EmptyState } from "./emptystate";
+export { EmptyState } from "./emptyState";
