+++ packages/raystack/v1/components/themeProvider/index.tsx
@@ -1,0 +1,4 @@
+export { ThemeSwitcher } from "./switcher";
+export { ThemeProvider, useTheme } from "./theme";
+export { ThemeProviderProps } from "./types";
+// Note: This themeProvider folder is a merge of old and the new themeProvider. Both old and the v1 folder contains the exact copy of themeprovider which was merged.
