--- packages/raystack/v1/components/themprovider/index.tsx
@@ -1,4 +1,0 @@
-export { ThemeSwitcher } from "./switcher";
-export { ThemeProvider, useTheme } from "./theme";
-export { ThemeProviderProps } from "./types";
-// Note: This themeprovider folder is a merge of old and the new themeprovider. Bothe old and the v1 folder contains the exact copy of themeprovider which was merged.
