+++ packages/raystack/v1/index.tsx
@@ -8,12 +8,13 @@ export { ToastContainer, toast } from "./components/toast";
 export { DropdownMenu } from "./components/dropdownMenu";
 export { Text } from "./components/text";
 export { Flex } from "./components/flex";
-export { EmptyState } from "./components/emptystate";
+export { EmptyState } from "./components/emptyState";
 export { Checkbox } from "./components/checkbox";
+export { Box } from "./components/box"
 
 export {
   ThemeProvider,
   ThemeProviderProps,
   ThemeSwitcher,
   useTheme,
-} from "./components/themprovider";
+} from "./components/themeProvider";
